feat(dsl): bring the Action playback verbs into @openmaic/dsl
Move the Action contract (the Action union + all variants, ActionType,
the FIRE_AND_FORGET_ACTIONS / SLIDE_ONLY_ACTIONS / SYNC_ACTIONS category
lists, and PercentageGeometry) from the app's lib/types/action.ts into a
new @openmaic/dsl/src/action.ts, keeping the package zero-runtime-dep.

The widget interaction actions (widget_highlight / widget_setState /
widget_annotation / widget_reveal) are config-free playback verbs after
the scene-action unification, so they graduate into the contract
alongside spotlight / laser / the wb_* family / discussion: the standard
Action union now covers them too.

Scene<TAction> now defaults to the standard Action union (was `never`);
skeleton-only consumers opt out with Scene<never, …>. lib/types/action.ts
collapses to a pure re-export shim, so every existing call site keeps
working unchanged. README/index/stage docs updated for the new runtime
layer in the dependency topology.

Refs #787 (Phase 2 Part A, under #720).
