+++ packages/@openmaic/dsl/src/stage.ts
@@ -3,10 +3,10 @@
  *
  * This module owns the *structural* part of a lesson: the top-level `Stage`
  * container and a per-page `Scene` whose `content` is a discriminated union of
- * the universal content kinds (`SlideContent`, `QuizContent`). Richer,
- * faster-moving feature surfaces (playback `Action`s, Ultra-mode widgets, PBL
- * project configs) are deliberately *not* defined here — they stay in the
- * consuming app and are threaded in through `Scene`'s generic parameters.
+ * the universal content kinds (`SlideContent`, `QuizContent`). `Scene`'s
+ * playback `actions` default to the contract's standard {@link Action} union
+ * (defined in `./action.ts`); apps still thread their own richer content
+ * kinds (interactive / PBL configs) in through `Scene`'s generic parameters.
  *
  * The split keeps `@openmaic/dsl` focused on the lesson skeleton while letting the
  * runtime engine, renderer, and importer share one source of truth for it.
@@ -14,6 +14,7 @@
  * No runtime dependencies. Pure types + pure discriminant guards only.
  */
 import type { Slide } from './slides.js';
+import type { Action } from './action.js';
 
 /** All scene kinds the contract is aware of. Feature kinds (interactive/pbl) are still valid `type` values — their *content* shapes live in the app and are composed in via {@link Scene}'s `TContent` parameter. */
 export type SceneType = 'slide' | 'quiz' | 'interactive' | 'pbl';
@@ -154,27 +155,31 @@ export type SceneContent = SlideContent | QuizContent;
 /**
  * Scene - Represents a single page/scene in the course.
  *
- * Generic so the contract owns only the universal skeleton while the app
- * injects its concrete playback action set and full content union:
+ * Generic so the contract owns the universal skeleton + the standard action
+ * set, while the app widens the content union (and, if needed, the action
+ * union) for its richer feature kinds:
  *
  * ```ts
- * // app side
+ * // app side — widen content; widen actions only if the app adds its own
  * type AppScene = Scene<Action, AppSceneContent>;
  * ```
  *
- * Defaults (`TAction = never`, `TContent = SlideContent | QuizContent`) yield a
- * read-only, feature-free scene — what renderers / importers that only care
- * about the skeleton want. The `TContent` constraint is structural — any union
- * of objects tagged with a `type: SceneType` discriminant satisfies it — so an
- * app can pass its own wider content union (slide | quiz | interactive | pbl).
+ * Defaults (`TAction = Action`, `TContent = SlideContent | QuizContent`) yield a
+ * scene whose `actions` are the contract's standard {@link Action} union and
+ * whose content spans the two universal kinds — what the runtime engine and
+ * playback consumers want out of the box. Skeleton-only consumers that reject
+ * actions entirely can still opt out with `Scene<never, …>`. The `TContent`
+ * constraint is structural — any union of objects tagged with a `type:
+ * SceneType` discriminant satisfies it — so an app can pass its own wider
+ * content union (slide | quiz | interactive | pbl).
  *
- * @template TAction  - The playback action type (defaults to `never`, i.e. none).
+ * @template TAction  - The playback action type (defaults to the standard {@link Action} union).
  * @template TContent - The scene-content union; any object union tagged with a
  *                      `type: {@link SceneType}` discriminant (defaults to the
  *                      two universal kinds).
  */
 export interface Scene<
-  TAction = never,
+  TAction = Action,
   TContent extends { type: SceneType } = SlideContent | QuizContent,
 > {
   id: string;
